refactor(matte): move Create Matte from toolbar to media panel overflow menu
A matte only adds an image to the library, so it belongs alongside
import/generate rather than the timeline editing toolbar. Extract MatteSheet
into its own file and present it from the media panel's overflow menu.
